requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
django-environ==0.8.1
idna==3.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
